+++ test/supabase_flutter_test.dart
@@ -6,6 +6,7 @@ import 'widget_test_stubs.dart';
 void main() {
   const supabaseUrl = '';
   const supabaseKey = '';
+
   group("Valid session", () {
     setUp(() async {
       mockAppLink();
@@ -14,6 +15,7 @@ void main() {
         url: supabaseUrl,
         anonKey: supabaseKey,
         localStorage: MockLocalStorage(),
+        pkceAsyncStorage: MockAsyncStorage(),
       );
     });
 
@@ -33,6 +35,7 @@ void main() {
         url: supabaseUrl,
         anonKey: supabaseKey,
         localStorage: MockLocalStorage(),
+        pkceAsyncStorage: MockAsyncStorage(),
       );
 
       final newClient = Supabase.instance.client;
@@ -47,11 +50,44 @@ void main() {
         url: supabaseUrl,
         anonKey: supabaseKey,
         localStorage: MockExpiredStorage(),
+        pkceAsyncStorage: MockAsyncStorage(),
       );
     });
+
+    tearDown(() => Supabase.instance.dispose());
+
     test('initial session contains the error', () async {
       await expectLater(
           SupabaseAuth.instance.initialSession, throwsA(isA<AuthException>()));
+    });
+  });
+
+  group('Deep Link with PKCE code', () {
+    late final PkceHttpClient pkceHttpClient;
+    setUp(() async {
+      pkceHttpClient = PkceHttpClient();
+
+      // Add initial deep link with a `code` parameter
+      mockAppLink(
+        initialLink: 'com.supabase://callback/?code=my-code-verifier',
+      );
+      await Supabase.initialize(
+        url: supabaseUrl,
+        anonKey: supabaseKey,
+        authFlowType: AuthFlowType.pkce,
+        httpClient: pkceHttpClient,
+        localStorage: MockEmptyLocalStorage(),
+        pkceAsyncStorage: MockAsyncStorage(),
+      );
+    });
+
+    tearDown(() => Supabase.instance.dispose());
+
+    test(
+        'Having `code` as the query parameter triggers `getSessionFromUrl` call on initialize',
+        () async {
+      expect(pkceHttpClient.requestCount, 1);
+      expect(pkceHttpClient.lastRequestBody['auth_code'], 'my-code-verifier');
     });
   });
 }