+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.9.0]
+
+- feat: add pkce support [#432](https://github.com/supabase/supabase-flutter/pull/432)
+  ```dart
+  // Initialize Supabase with AuthFlowType.pkce to enable PKCE flow for deep link related auth
+  await Supabase.initialize(
+    url: supabaseUrl,
+    anonKey: supabaseKey,
+    authFlowType: AuthFlowType.pkce,
+  );
+  ```
+
 ## [1.8.1]
 
 - fix: ensure that Google login on Android always opens in external browser [#455](https://github.com/supabase/supabase-flutter/pull/455)